# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF677 - Usage Consumption\n\n### Release: 18.5\
    \ - December 2018\n\nThis API covers the consumption follow up function providing\
    \ ongoing information about usages related to any subscribed communication products\
    \ (voice, data, TV,…) without having to wait the invoice production. These information\
    \ may concern usages charged on a bucket supervised or not and so the remaining\
    \ credits on the bucket. It allows customers or users to be informed on usages\
    \ done and remaining credits on the buckets that they consume under their purchased\
    \ offers and options.\n\nThe usage consumption API allow to view at a given point\
    \ the balance and the consumption counters of the various buckets (SMS, Voice,\
    \ Data for example) that one or more user(s) consume with each of his devices,\
    \ according to the purchased offers and options.\n\n### Resources \n- **usageConsumptionReport**\
    \ : an usage consumption report retrieves the data related to these balances and\
    \ various consumption counters and calculated at the time of the request by the\
    \ server.\n- **usageConsumptionReportRequest** : used to manage the calculation\
    \ request of an usage consumption report.\n\n### Operations\nThe use of the both\
    \ resources described in this document is conditioned by the capability of the\
...
    \ without necessarily store it.\n- We must also anticipate the use case where\
...
host: "localhost:8080"
basePath: "/tmf-api"
...
schemes:
...
consumes:
...
produces:
...
paths:
...
  /usageConsumptionReportRequest:
    get:
      tags:
      - "usageConsumptionReportRequest"
      summary: "List or find UsageConsumptionReportRequest objects"
      description: "This operation list or find UsageConsumptionReportRequest entities"
      operationId: "listUsageConsumptionReportRequest"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/UsageConsumptionReportRequest"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "UsageConsumptionReportRequest"
    post:
      tags:
      - "usageConsumptionReportRequest"
      summary: "Creates a UsageConsumptionReportRequest"
      description: "This operation creates a UsageConsumptionReportRequest entity."
      operationId: "createUsageConsumptionReportRequest"
      parameters:
      - in: "body"
        name: "usageConsumptionReportRequest"
        description: "The UsageConsumptionReportRequest to be created"
        required: true
        schema:
          $ref: "#/definitions/UsageConsumptionReportRequest_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/UsageConsumptionReportRequest"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "UsageConsumptionReportRequest"
  /usageConsumptionReportRequest/{id}:
    get:
      tags:
      - "usageConsumptionReportRequest"
      summary: "Retrieves a UsageConsumptionReportRequest by ID"
      description: "This operation retrieves a UsageConsumptionReportRequest entity.\
        \ Attribute selection is enabled for all first level attributes."
      operationId: "retrieveUsageConsumptionReportRequest"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the UsageConsumptionReportRequest"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/UsageConsumptionReportRequest"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "UsageConsumptionReportRequest"
    delete:
      tags:
      - "usageConsumptionReportRequest"
      summary: "Deletes a UsageConsumptionReportRequest"
      description: "This operation deletes a UsageConsumptionReportRequest entity."
      operationId: "deleteUsageConsumptionReportRequest"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the UsageConsumptionReportRequest"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "UsageConsumptionReportRequest"
...
definitions:
...
  NetworkProductRef:
    type: "object"
    required:
    - "href"
    - "id"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the network product"
      href:
        type: "string"
        description: "Reference to the network product"
      name:
        type: "string"
        description: "Network product name"
      publicIdentifier:
        type: "string"
        description: "Public number associated to the product (msisdn number for mobile\
          \ line for example)"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "Generic attribute indicating the name of the class type of the\
          \ referred resource entity"
    description: "Reference of a product"
    example:
      '@referredType': "@referredType"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
      publicIdentifier: "publicIdentifier"
...
  RelatedParty:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of a related party"
      href:
        type: "string"
        description: "Reference of the related party, could be a party reference or\
          \ a party role reference"
      name:
        type: "string"
        description: "Name of the related party"
      role:
        type: "string"
        description: "Role of the related party."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "The actual type of the target instance when needed for disambiguation."
    description: "RelatedParty reference. A related party defines party or party role\
      \ linked to a specific entity."
    example:
      '@referredType': "@referredType"
      role: "role"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "@schemaLocation"
  TimePeriod:
    type: "object"
    properties:
      endDateTime:
        type: "string"
        format: "date-time"
        description: "End of the time period, using IETC-RFC-3339 format"
      startDateTime:
        type: "string"
        format: "date-time"
        description: "Start of the time period, using IETC-RFC-3339 format. If you\
          \ define a start, you must also define an end"
    description: "A period of time, either as a deadline (endDateTime only) a startDateTime\
      \ only, or both"
    example:
      startDateTime: "2000-01-23T04:56:07.000+00:00"
      endDateTime: "2000-01-23T04:56:07.000+00:00"
...
  UsageConsumptionReportRef:
    type: "object"
    required:
    - "href"
    - "id"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the usage consumption report given by the\
          \ server"
      href:
        type: "string"
        description: "Hyperlink to access the usage consumption report"
      description:
        type: "string"
        description: "Free short text describing the usage consumption report content"
      effectiveDate:
        type: "string"
        format: "date-time"
        description: "Date and time when the usage consumption report was calculated\
          \ and generated"
      name:
        type: "string"
        description: "Usage consumption report name"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        format: "uri"
        description: "The type of the referred entity"
    description: "Reference of an usage consumption report"
    example:
      '@referredType': "http://example.com/aeiou"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      description: "description"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
      effectiveDate: "2000-01-23T04:56:07.000+00:00"
  UsageConsumptionReportRequest:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the usage consumption report request given\
          \ by the server"
      href:
        type: "string"
        description: "Reference of the usage consumption report request"
      creationDate:
        type: "string"
        format: "date-time"
        description: "Date and time of the request creation"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date when the status was last changed"
      status:
        type: "string"
        description: "Status of the usage consumption report request (InProgress or\
          \ done)"
      bucket:
        type: "array"
        description: "Reference of the buckets for which the usage consumption report\
          \ is requested"
        items:
          $ref: "#/definitions/UsageVolumeProductRef"
      product:
        description: "Network product reference for which the usage consumption report\
          \ is requested"
        $ref: "#/definitions/NetworkProductRef"
      relatedParty:
        type: "array"
        description: "Reference and role of the related parties for which the usage\
          \ consumption report is requested"
        items:
          $ref: "#/definitions/RelatedParty"
      usageConsumptionReport:
        description: "References of the usage consumption report (given when it has\
          \ been calculated and the status of the request is done)"
        $ref: "#/definitions/UsageConsumptionReportRef"
      validPeriod:
        description: "Validity period"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "An UsageConsumptionReportRequest allows to manage the calculation\
      \ request of an usage consumption report"
    example:
      product:
        '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        publicIdentifier: "publicIdentifier"
      usageConsumptionReport:
        '@referredType': "http://example.com/aeiou"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        description: "description"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        effectiveDate: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      creationDate: "2000-01-23T04:56:07.000+00:00"
      relatedParty:
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      validPeriod:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      bucket:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        '@schemaLocation': "http://example.com/aeiou"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        '@schemaLocation': "http://example.com/aeiou"
      '@baseType': "@baseType"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
      status: "status"
  UsageConsumptionReportRequest_Create:
    type: "object"
    properties:
      creationDate:
        type: "string"
        format: "date-time"
        description: "Date and time of the request creation"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date when the status was last changed"
      status:
        type: "string"
        description: "Status of the usage consumption report request (InProgress or\
          \ done)"
      bucket:
        type: "array"
        description: "Reference of the buckets for which the usage consumption report\
          \ is requested"
        items:
          $ref: "#/definitions/UsageVolumeProductRef"
      product:
        description: "Network product reference for which the usage consumption report\
          \ is requested"
        $ref: "#/definitions/NetworkProductRef"
      relatedParty:
        type: "array"
        description: "Reference and role of the related parties for which the usage\
          \ consumption report is requested"
        items:
          $ref: "#/definitions/RelatedParty"
      usageConsumptionReport:
        description: "References of the usage consumption report (given when it has\
          \ been calculated and the status of the request is done)"
        $ref: "#/definitions/UsageConsumptionReportRef"
      validPeriod:
        description: "Validity period"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "An UsageConsumptionReportRequest allows to manage the calculation\
      \ request of an usage consumption report\nSkipped properties: id,href"
    example:
      bucket:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        '@schemaLocation': "http://example.com/aeiou"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        '@schemaLocation': "http://example.com/aeiou"
      product:
        '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        publicIdentifier: "publicIdentifier"
      usageConsumptionReport:
        '@referredType': "http://example.com/aeiou"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        description: "description"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        effectiveDate: "2000-01-23T04:56:07.000+00:00"
      '@baseType': "@baseType"
      '@type': "@type"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      creationDate: "2000-01-23T04:56:07.000+00:00"
      relatedParty:
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      '@schemaLocation': "http://example.com/aeiou"
      validPeriod:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      status: "status"
...
  UsageVolumeProductRef:
    type: "object"
    required:
    - "id"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the bucket"
      name:
        type: "string"
        description: "Bucket name"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "Generic attribute indicating the name of the class type of the\
          \ referred resource entity"
    description: "Reference of a bucket"
    example:
      '@referredType': "@referredType"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      id: "id"
      '@schemaLocation': "http://example.com/aeiou"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"